internal/cmd/gocdk: factor out launcher code to new package
Not changing API yet, that will come in a follow-up PR.

Updates #2046
